+++ python/packages/nisar/products/utils.py
@@ -76,3 +76,48 @@ def get_static_layers_data_access(
                                                        granule_id)
 
     return static_layers_data_access
+
+
+def interpret_subswath_mask(subswath_mask, nodata=255):
+    """
+    Interprets a subswath mask integer by decoding its digits into boolean 
+    flags indicating reference validity, secondary validity, and water
+    presence.
+
+    Parameters
+    ----------
+    subswath_mask : numpy.array
+        Each digit represents a specific flag:
+        - Units digit (1s place): Secondary subswath mask
+            Non-zero indicates valid; zero indicates invalid.
+        - Tens digit (10s place): Reference subswath mask
+            Non-zero indicates valid; zero indicates invalid.
+        - Hundreds digit (100s place): Water presence flag.
+            Non-zero indicates presence of water; zero indicates absence.
+    nodata : int, default 255
+
+    Returns
+    -------
+    reference_valid : bool
+        True if the reference is valid (tens digit is non-zero),
+        False otherwise.
+    secondary_valid : bool
+        True if the secondary is valid (units digit is non-zero),
+        False otherwise.
+    water : bool
+        True if water is present (hundreds digit is non-zero),
+        False otherwise.
+    """
+    arr = np.asarray(subswath_mask)
+
+    nd = (arr == nodata)
+
+    secondary_valid = subswath_mask % 10 != 0
+    reference_valid = (subswath_mask // 10) % 10 != 0
+    water = (subswath_mask // 100) % 10 != 0
+
+    secondary_valid = np.where(nd, False, secondary_valid)
+    reference_valid = np.where(nd, False, reference_valid)
+    water = np.where(nd, False, water)
+
+    return reference_valid, secondary_valid, water
